+++ crates/test-runner/src/inspectors/mod.rs
@@ -1,2 +1,0 @@
-/// The Cheatcodes inspector
-pub mod cheats_inspector;