+++ crates/utils/src/shell.rs
@@ -1,0 +1,2 @@
+/// The verbosity level
+pub type Verbosity = u8;